+++ test/private/outlineMembers.test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { throws } = require('assert');
+const { strictEqual, throws } = require('assert');
 const { resolve } = require('path');
 const { stringify } = require('flatted');
 const snapshot = require('snapshot-assertion');
@@ -14,9 +14,20 @@ module.exports = (tests) => {
   tests.add(
     '`outlineMembers` with first argument `members` not an array.',
     () => {
+      const codeFiles = new Map([[TEST_CODE_FILE_PATH, '']]);
+
       throws(() => {
-        outlineMembers(true);
+        outlineMembers(true, codeFiles);
       }, new TypeError('First argument “members” must be an array.'));
+    }
+  );
+
+  tests.add(
+    '`outlineMembers` with second argument `codeFiles` not a Map instance.',
+    () => {
+      throws(() => {
+        outlineMembers([], true);
+      }, new TypeError('Second argument “codeFiles” must be a Map instance.'));
     }
   );
 
@@ -79,16 +90,26 @@ module.exports = (tests) => {
     );
 
     await snapshot(
-      stringify(outlineMembers(members), null, 2),
-      resolve(__dirname, '../snapshots/outlineMembers.json')
+      stringify(outlineMembers(members, codeFiles), null, 2),
+      resolve(
+        __dirname,
+        '../snapshots/outlineMembers/no-missing-jsdoc-members.json'
+      )
     );
   });
 
-  tests.add('`outlineMembers` with missing members.', async () => {
-    const code = `/**
- * @kind member
- * @name A.a
- */`;
+  tests.add('`outlineMembers` with a missing member.', async () => {
+    const code = `// Code before…
+
+class A {
+  /**
+   * @kind function
+   * @name A#a
+   */
+  a() {}
+}
+
+// Code after…`;
     const codeFiles = new Map([[TEST_CODE_FILE_PATH, code]]);
     const jsdocComments = await codeToJsdocComments(code, TEST_CODE_FILE_PATH);
     const members = jsdocCommentsToMembers(
@@ -97,8 +118,22 @@ module.exports = (tests) => {
       TEST_CODE_FILE_PATH
     );
 
-    throws(() => {
-      outlineMembers(members);
-    }, new Error('Missing JSDoc for namepath “A”.'));
+    let caughtError;
+
+    try {
+      outlineMembers(members, codeFiles);
+    } catch (error) {
+      caughtError = error;
+    }
+
+    strictEqual(caughtError instanceof Error, true);
+
+    await snapshot(
+      caughtError.message,
+      resolve(
+        __dirname,
+        '../snapshots/outlineMembers/error-namepath-missing-jsdoc-member.ans'
+      )
+    );
   });
 };