+++ changelog.md
@@ -17,7 +17,7 @@
   - `@property`
   - `@return`
   - `@typedef`
-- Display the relevant file path and location in the source code for JSDoc namepath syntax error messages, using a new [`@babel/code-frame`](https://npm.im/@babel/code-frame) dependency.
+- Display the relevant source code file path and location in JSDoc namepath related error messages, using a new [`@babel/code-frame`](https://npm.im/@babel/code-frame) dependency.
 
 ### Patch